automatically backed up by learn

+++ index.js
@@ -53,7 +53,7 @@ function makeInt(n) {
 makeInt()
 
 function preserveDecimal(n) {
-  return parseInt(n)
+  return parseFloat(n)
 }
 
 preserveDecimal()
